Rework for intangible assets. Initial commit, more to come
