+++ src/main/java/uk/gov/companieshouse/api/accounts/ResourceName.java
@@ -14,7 +14,7 @@ public enum ResourceName {
     CREDITORS_AFTER_ONE_YEAR("creditors-after-more-than-one-year"),
     STOCKS("stocks"),
     TANGIBLE_ASSETS("tangible-assets"),
-    INTANGIBLE_ASSETS("intangible-assets"),
+    INTANGIBLE_ASSETS("intangibleassets-assets"),
     CURRENT_ASSETS_INVESTMENTS("current-assets-investments"),
     FIXED_ASSETS_INVESTMENTS("fixed-assets-investments"),
     EMPLOYEES("employees"),
