+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/smallfull/notes/intangibleassets/CostEntity.java
@@ -1,38 +1,25 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.intangible;
+package uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.intangibleassets;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import org.hibernate.validator.constraints.Range;
+import org.springframework.data.mongodb.core.mapping.Field;
 
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class Cost {
+public class CostEntity {
 
-    private static final int MAX_RANGE = 99999999;
-    private static final int ZERO = 0;
-    private static final int MIN_RANGE = -99999999;
-
-    @Range(min = ZERO, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("additions")
+    @Field("additions")
     private Long additions;
 
-    @Range(min = ZERO, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("at_period_end")
+    @Field("at_period_end")
     private Long atPeriodEnd;
 
-    @Range(min = ZERO, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("at_period_start")
+    @Field("at_period_start")
     private Long atPeriodStart;
 
-    @Range(min = ZERO, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("disposals")
+    @Field("disposals")
     private Long disposals;
 
-    @Range(min = MIN_RANGE, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("revaluations")
+    @Field("revaluations")
     private Long revaluations;
 
-    @Range(min = MIN_RANGE, max = MAX_RANGE, message = "value.outside.range")
-    @JsonProperty("transfers")
+    @Field("transfers")
     private Long transfers;
 
     public Long getAdditions() {
@@ -81,5 +68,17 @@ public class Cost {
 
     public void setTransfers(Long transfers) {
         this.transfers = transfers;
+    }
+
+    @Override
+    public String toString() {
+        return "CostEntity{" +
+                "additions=" + additions +
+                ", atPeriodEnd=" + atPeriodEnd +
+                ", atPeriodStart=" + atPeriodStart +
+                ", disposals=" + disposals +
+                ", revaluations=" + revaluations +
+                ", transfers=" + transfers +
+                '}';
     }
 }
