--- src/main/java/uk/gov/companieshouse/api/accounts/repository/IntangibleAssetsRepository.java
@@ -1,10 +1,0 @@
-package uk.gov.companieshouse.api.accounts.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.intangible.IntangibleAssetsEntity;
-
-@Repository
-public interface IntangibleAssetsRepository extends MongoRepository<IntangibleAssetsEntity, String>{
-
-}
