+++ src/main/java/uk/gov/companieshouse/api/accounts/repository/smallfull/IntangibleAssetsRepository.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.api.accounts.repository.smallfull;
+
+import org.springframework.stereotype.Repository;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.intangibleassets.IntangibleAssetsEntity;
+import uk.gov.companieshouse.api.accounts.repository.AccountsNoteRepository;
+
+@Repository
+public interface IntangibleAssetsRepository extends AccountsNoteRepository<IntangibleAssetsEntity> {
+
+    @Override
+    default AccountingNoteType getAccountsNote() {
+        return AccountingNoteType.SMALL_FULL_INTANGIBLE_ASSETS;
+    }
+}
