+++ src/main/resources/notes.yml
@@ -6,3 +6,4 @@ notes:
     - employees
     - debtors
     - current-assets-investments
+    - intangible-assets
